refactor: operation 패키지 -> game 패키지로 이름 수정
