+++ src/main/java/lotto/common/config/AppConfig.java
@@ -3,8 +3,8 @@ package lotto.common.config;
 import lotto.controller.LottoController;
 import lotto.common.handler.InputHandler;
 import lotto.common.handler.OutputHandler;
-import lotto.model.domain.operation.NumberGenerator;
-import lotto.model.domain.operation.RandomNumberGenerator;
+import lotto.model.domain.game.NumberGenerator;
+import lotto.model.domain.game.RandomNumberGenerator;
 import lotto.model.service.LottoService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
