+++ src/main/java/lotto/model/domain/game/WinningStatistics.java
@@ -1,4 +1,4 @@
-package lotto.model.domain.operation;
+package lotto.model.domain.game;
 
 import java.util.EnumMap;
 import java.util.Map.Entry;
