+++ src/test/java/lotto/model/domain/LottosTest.java
@@ -3,7 +3,7 @@ package lotto.model.domain;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.List;
-import lotto.model.domain.operation.NumberGenerator;
+import lotto.model.domain.game.NumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
